requirements.txt
cached-property==1.5.2
certifi==2020.11.8
cycler==0.10.0
h5py==3.1.0
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.2.0
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.5.4
six==1.15.0
slayerCuda==0.0.0
slayerLoihiCuda==0.0.0
strictyaml==1.1.0
tensorboardX==2.1
torch==1.4.0
torchsparse==1.1.0
torchvision==0.5.0
tqdm==4.51.0
